Improve code duplication
