+++ src/index.ts
@@ -12,141 +12,9 @@ import { Dataset } from "./types/api.js";
 import { HoneycombError } from "./utils/errors.js";
 import process from "node:process";
 import { AnalysisQuery } from "./types/query.js";
-
-/**
- * Calculate summary statistics for query results to provide useful insights
- * without overwhelming the context window
- */
-function summarizeResults(results: any[], params: z.infer<typeof QueryToolSchema>) {
-  if (!results || results.length === 0) {
-    return { count: 0 };
-  }
-  
-  const summary = {
-    count: results.length,
-  };
-  
-  // If we have calculation columns, add some statistics about them
-  if (params.calculations) {
-    // Get numeric calculation columns (excluding HEATMAP, COUNT, CONCURRENCY)
-    const numericColumns = params.calculations
-      .filter(calc => 
-        calc.op !== "COUNT" && 
-        calc.op !== "CONCURRENCY" && 
-        calc.op !== "HEATMAP" &&
-        calc.column
-      )
-      .map(calc => `${calc.op}(${calc.column})`);
-    
-    // Add summary stats for each numeric column
-    numericColumns.forEach(colName => {
-      if (results[0] && colName in results[0]) {
-        // Calculate min/max/avg if the column exists in results
-        const values = results.map(r => r[colName]).filter(v => v !== null && v !== undefined);
-        if (values.length > 0) {
-          const min = Math.min(...values);
-          const max = Math.max(...values);
-          const sum = values.reduce((a, b) => a + b, 0);
-          const avg = sum / values.length;
-          
-          // Calculate median (P50 approximation)
-          const sortedValues = [...values].sort((a, b) => a - b);
-          const medianIndex = Math.floor(sortedValues.length / 2);
-          const median = sortedValues.length % 2 === 0
-            ? (sortedValues[medianIndex - 1] + sortedValues[medianIndex]) / 2
-            : sortedValues[medianIndex];
-          
-          // @ts-ignore - dynamically adding properties
-          summary[colName] = { 
-            min, 
-            max, 
-            avg,
-            median,
-            sum,
-            // Add range and standard deviation for better distribution insight
-            range: max - min,
-            stdDev: calculateStdDev(values, avg)
-          };
-        }
-      }
-    });
-    
-    // Special handling for COUNT operations
-    const hasCount = params.calculations.some(calc => calc.op === "COUNT");
-    if (hasCount && results.length > 0 && 'COUNT' in results[0]) {
-      const countValues = results.map(r => r.COUNT).filter(v => v !== null && v !== undefined);
-      if (countValues.length > 0) {
-        const totalCount = countValues.reduce((a, b) => a + b, 0);
-        const maxCount = Math.max(...countValues);
-        const minCount = Math.min(...countValues);
-        
-        // @ts-ignore - dynamically adding properties
-        summary.countStats = {
-          total: totalCount,
-          max: maxCount, 
-          min: minCount,
-          avg: totalCount / countValues.length
-        };
-      }
-    }
-  }
-  
-  // Add unique count for breakdown columns
-  if (params.breakdowns && params.breakdowns.length > 0) {
-    const breakdownStats = {};
-    
-    params.breakdowns.forEach(col => {
-      const uniqueValues = new Set(results.map(r => r[col]));
-      // @ts-ignore - dynamically adding properties
-      breakdownStats[col] = {
-        uniqueCount: uniqueValues.size,
-        // Add top values (up to 5) with their frequencies
-        topValues: getTopValues(results, col, 5)
-      };
-    });
-    
-    // @ts-ignore - dynamically adding properties
-    summary.breakdowns = breakdownStats;
-  }
-  
-  return summary;
-}
-
-/**
- * Calculate standard deviation for an array of values
- */
-function calculateStdDev(values: number[], mean: number): number {
-  if (values.length <= 1) return 0;
-  
-  const squareDiffs = values.map(value => {
-    const diff = value - mean;
-    return diff * diff;
-  });
-  
-  const avgSquareDiff = squareDiffs.reduce((sum, value) => sum + value, 0) / squareDiffs.length;
-  return Math.sqrt(avgSquareDiff);
-}
-
-/**
- * Get top N values with their frequencies
- */
-function getTopValues(results: any[], column: string, limit: number = 5): Array<{value: any, count: number}> {
-  const valueCounts = new Map();
-  
-  // Count frequencies
-  results.forEach(result => {
-    const value = result[column];
-    if (value !== undefined && value !== null) {
-      valueCounts.set(value, (valueCounts.get(value) || 0) + 1);
-    }
-  });
-  
-  // Convert to array and sort by frequency
-  return Array.from(valueCounts.entries())
-    .map(([value, count]) => ({ value, count }))
-    .sort((a, b) => b.count - a.count)
-    .slice(0, limit);
-}
+import { calculateStdDev, getTopValues } from "./utils/functions.js";
+import { summarizeResults } from "./transformations.js";
+import { generateInterpretation } from "./utils/analysis.js";
 
 async function handleToolError(
   error: unknown,
@@ -701,37 +569,4 @@ if (import.meta.url === `file://${process.argv[1]}`) {
     console.error("Fatal error in main():", error);
     process.exit(1);
   });
-}
-
-/**
- * Generate interpretation text for numeric statistics based on Honeycomb documentation
- */
-function generateInterpretation(stats: {
-  min?: number;
-  max?: number;
-  avg?: number;
-  p95?: number;
-  [key: string]: any
-}, columnName: string): string {
-  const interpretations = [];
-  
-  if (stats.avg !== undefined && stats.p95 !== undefined) {
-    const ratio = stats.p95 / stats.avg;
-    if (ratio > 3) {
-      interpretations.push(`The P95 value is ${ratio.toFixed(1)}x higher than the average, suggesting significant outliers in ${columnName}.`);
-    }
-  }
-  
-  if (stats.min !== undefined && stats.max !== undefined) {
-    const range = stats.max - stats.min;
-    if (stats.avg !== undefined && range > stats.avg * 10) {
-      interpretations.push(`The range (${range}) is very wide compared to the average (${stats.avg}), indicating high variability.`);
-    }
-  }
-  
-  if (interpretations.length === 0) {
-    return `Standard distribution of ${columnName} values with expected statistical properties.`;
-  }
-  
-  return interpretations.join(' ');
 }
